used parcel, installed tailwindcss and converted to ES module
