+++ index.html
@@ -33,6 +33,6 @@
       />
       <div id="movieList" class="grid grid-cols-2 md:grid-cols-4 gap-4"></div>
     </div>
-    <script src="./main.js"></script>
+    <script type="module" src="./src/modules/main.js"></script>
   </body>
 </html>
