+++ journal.html
@@ -28,7 +28,6 @@
       <h1 class="text-3xl font-bold mb-4">Favorite Movies</h1>
       <div id="favMovies" class="grid grid-cols-2 md:grid-cols-4 gap-4"></div>
     </div>
-    <script src="./journal.js"></script>
-    <script src="./main.js"></script>
+    <script type="module" src="./src/modules/journal.js"></script>
   </body>
 </html>
