+++ src/modules/main.js
@@ -1,12 +1,12 @@
+// src/modules/main.js
+import { addToFavorite } from "./storage.js";
+
 async function fetchMovies() {
   const apiUrl =
     "https://api.themoviedb.org/3/movie/popular?api_key=d4f9e0502356d1d0ce1ae299c5217299";
 
   try {
     const response = await fetch(apiUrl);
-    if (!response.ok) {
-      throw new Error(`HTTP error! Status: ${response.status}`);
-    }
     const data = await response.json();
     renderMovies(data.results);
     setupSearch(data.results);
@@ -17,19 +17,11 @@ async function fetchMovies() {
 
 function setupSearch(movies) {
   const searchInput = document.getElementById("searchInput");
-
   searchInput.addEventListener("input", (event) => {
     const searchTerm = event.target.value.toLowerCase();
-
-    if (searchTerm === "") {
-      renderMovies(movies);
-      return;
-    }
-
     const filteredMovies = movies.filter((movie) =>
       movie.title.toLowerCase().includes(searchTerm)
     );
-
     renderMovies(filteredMovies);
   });
 }
@@ -40,17 +32,11 @@ function renderMovies(movies) {
 
   movies.forEach((movie) => {
     const movieCard = document.createElement("div");
-    movieCard.classList.add("bg-white", "rounded-lg", "shadow-md", "p-4");
-
+    movieCard.className = "bg-white rounded-lg shadow-md p-4";
     movieCard.innerHTML = `
-      <img class="w-full h-[300px] object-cover rounded" src="https://image.tmdb.org/t/p/w500${
-        movie.poster_path
-      }" alt="${movie.title}" />
+      <img class="w-full h-[300px] object-cover rounded" src="https://image.tmdb.org/t/p/w500${movie.poster_path}" alt="${movie.title}" />
       <h2 class="mt-2 text-lg font-semibold">${movie.title}</h2>
-      <p class="text-sm text-gray-600">${movie.overview.substring(
-        0,
-        100
-      )}...</p>
+      <p class="text-sm text-gray-600">${movie.overview.substring(0, 100)}...</p>
       <button class="mt-2 bg-gray-500 text-white px-3 py-1 rounded hover:bg-blue-600 add-fav-btn">
         Add to Favorites
       </button>
@@ -61,23 +47,6 @@ function renderMovies(movies) {
 
     movieList.appendChild(movieCard);
   });
-}
-
-const favMovies = JSON.parse(localStorage.getItem("favMoviesList")) || [];
-
-function addToFavorite(movie) {
-  const isAlreadyExisting = favMovies.some((mov) => mov.id === movie.id);
-  if (!isAlreadyExisting) {
-    favMovies.unshift(movie);
-    saveToStorage();
-    alert(`${movie.title} added to favorites.`);
-  } else {
-    alert(`${movie.title} is already in favorites.`);
-  }
-}
-
-function saveToStorage() {
-  localStorage.setItem("favMoviesList", JSON.stringify(favMovies));
 }
 
 fetchMovies();
